refactor: 컴포넌트 디렉터리 이동 및 타입 파일 분리
