+++ src/customPages/RoadmapEditKeywordPage/components/EditKeywordModal/type.ts
@@ -1,0 +1,9 @@
+import { ChildrenKeyword } from '../../../../types';
+
+export type EditKeywordModalProps = {
+  open: boolean;
+  onClose: () => void;
+  keywordContents: ChildrenKeyword;
+  sessionId: number;
+  keywordId: number;
+};
