+++ src/customPages/RoadmapEditKeywordPage/components/SubKeywordList/type.ts
@@ -1,0 +1,6 @@
+import { ChildrenKeyword } from '../../../../types';
+
+export type SubKeywordListProps = {
+  childrenKeywordList: ChildrenKeyword[];
+  sessionId: number;
+};
