+++ src/customPages/RoadmapSelectSessionPage/components/SessionList/index.tsx
@@ -6,22 +6,9 @@ import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import Paper from '@mui/material/Paper';
 import Button from '@mui/material/Button';
+import { SessionListProps } from './type';
 
-
-type ListProps = {
-  rows: { id: number }[];
-  onClickMove: (id: number) => void;
-  // onClickEdit;
-  // onClickDelete;
-};
-
-const SessionList = ({
-  rows,
-  onClickMove,
-}: // onClickAdd,
-// onClickEdit,
-// onClickDelete,
-ListProps) => {
+const SessionList = ({ rows, onClickMove }: SessionListProps) => {
   const columns = Object.keys(rows[0]);
 
   return (
