+++ src/customPages/RoadmapSelectSessionPage/components/SessionList/type.ts
@@ -1,0 +1,4 @@
+export type SessionListProps = {
+  rows: { id: number }[];
+  onClickMove: (id: number) => void;
+};
