+++ src/customPages/RoadmapSelectTopKeywordPage/components/TopKeywordList/type.ts
@@ -1,0 +1,4 @@
+export type TopKeywordListProps = {
+  rows: { keywordId: number }[];
+  onClickMove: (keywordId: number) => void;
+};
